% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-erp-create-matrices.R
\name{create_corr_ts}
\alias{create_corr_ts}
\title{Title}
\usage{
create_corr_ts(inp_cormat, cat_idx, time_ms_idx, abs = TRUE, fisher = TRUE,
  plot_col_range = c(-0.2, 2.5), smoothed = FALSE, diff = FALSE)
}
\arguments{
\item{diff}{}
}
\description{
Title
}
