% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-erp-create-matrices.R
\name{create_crplot_name}
\alias{create_crplot_name}
\title{Title}
\usage{
create_crplot_name(cat_idx, abs, fisher, time_val, file_ext = ".png")
}
\arguments{
\item{file_ext}{}
}
\description{
Title
}
