% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-erp-create-matrices.R
\name{create_crplot_title}
\alias{create_crplot_title}
\title{Title}
\usage{
create_crplot_title(cat_idx, abs, fisher, time_val, smoothed = FALSE)
}
\arguments{
\item{smoothed}{}
}
\description{
Title
}
