% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-read-erp.R
\name{detect_pattern}
\alias{detect_pattern}
\title{Detect whether a regex pattern exists in a given input string}
\usage{
detect_pattern(inp_str, str_pattern, ignore_case = TRUE)
}
\arguments{
\item{inp_str}{The input string in which to detect a pattern}

\item{str_pattern}{The regex pattern to detect in the input string}

\item{ignore_case}{Logical. When \code{TRUE} and in an interactive session case
will be ignored in the regex pattern detection. This is the default.}
}
\value{
Logical. \code{TRUE} if the pattern is detected in the string
}
\description{
Detect whether a regex pattern exists in a given input string
}
