% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-erp-create-matrices.R
\name{diff_mat_threshold}
\alias{diff_mat_threshold}
\title{Threshold values of a matrix based on upper and lower thresholds}
\usage{
diff_mat_threshold(inp_mat, lower_th = -0.9, upper_thr = 0.9)
}
\arguments{
\item{upper_thr}{}
}
\description{
Threshold values of a matrix based on upper and lower thresholds
}
