% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-erp-create-matrices.R
\name{erp_create_cat_all_mats}
\alias{erp_create_cat_all_mats}
\title{Title}
\usage{
erp_create_cat_all_mats(erp_df_labeled, core_cat, all_erp_time_int, bin_width)
}
\arguments{
\item{bin_width}{}
}
\description{
Title
}
