% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-erp-create-matrices.R
\name{erp_create_df_labeled}
\alias{erp_create_df_labeled}
\title{Title}
\usage{
erp_create_df_labeled(n_data_type, n_patient_num, session_index, task, filter,
  channel_index)
}
\arguments{
\item{channel_index}{}
}
\description{
Title
}
