% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-erp-spectrogram.R
\name{erp_create_specgram_trial}
\alias{erp_create_specgram_trial}
\title{Creates a spectrogram from the univariate time series x
This function is created using code from Peter Elliot (CMU Statistics)}
\usage{
erp_create_specgram_trial(inp_trial, Fs = 1000, n = 128, overlap = 96)
}
\arguments{
\item{n}{is the number of time points to use per Fourier transform}

\item{overlap}{overlap controls the jump from bin to bin the time resolution
of the spectrogram is (n - overlap)}

\item{is}{the sampling rate (samples per second)}
}
\value{
s is a list with three values
s$t is a vector of times associated with the beginning of each bin
s$f is a vector of frequencies that are measured for each bin.
s$f spans 0 to Fs/2 and has length n/2
s$S is a matrix of intensities - Frequency x Time
}
\description{
Creates a spectrogram from the univariate time series x
This function is created using code from Peter Elliot (CMU Statistics)
}
