% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-erp-spectrogram.R
\name{erp_plot_specgram_trial}
\alias{erp_plot_specgram_trial}
\title{Create and store the spectrogram plot using a customized \code{copper} color
palette
This function is created using code from Peter Elliot (CMU Statistics)}
\usage{
erp_plot_specgram_trial(inp_sgram)
}
\arguments{
\item{inp_sgram}{The input spectrogram object as created from the
\code{signal::specgram} function}
}
\description{
Create and store the spectrogram plot using a customized \code{copper} color
palette
This function is created using code from Peter Elliot (CMU Statistics)
}
