% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-erp-create-matrices.R
\name{erp_rbind_sessions}
\alias{erp_rbind_sessions}
\title{Rbind session data frames across 2 lists containing dataframes
for each individual session}
\usage{
erp_rbind_sessions(erp_df_lab1, erp_df_lab2)
}
\arguments{
\item{erp_df_lab2}{}
}
\description{
Rbind session data frames across 2 lists containing dataframes
for each individual session
}
