% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-erp-create-matrices.R
\name{erp_save_plot}
\alias{erp_save_plot}
\title{Title}
\usage{
erp_save_plot(time_idx, n_patient_num, batch_idx, cat_idx, session_idx,
  smoothed, out_dir, diff_type = "int", abs = TRUE, fisher = TRUE,
  plot_gg, thr_type = NULL)
}
\arguments{
\item{thr_type}{}
}
\description{
Title
}
