% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-read-erp.R
\name{ext_repl_pattern}
\alias{ext_repl_pattern}
\title{Extract text and replace it based on a specified pattern}
\usage{
ext_repl_pattern(inp_str, ext_pat, sub_pat, repl_pat = "",
  ignore_case = TRUE)
}
\arguments{
\item{inp_str}{The input string in which to extract a pattern}

\item{ext_pat}{The regex pattern to extract from the input string}

\item{sub_pat}{The regex pattern to substitute in the extracted string}

\item{repl_pat}{The replacement string to substitute in the extracted string}

\item{ignore_case}{Logical. When \code{TRUE} and in an interactive session case
will be ignored in the regex pattern detection. This is the default.}
}
\value{
Character. The extracted string with the substituted pattern
}
\description{
Extract text and replace it based on a specified pattern
}
