% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-erp-create-matrices.R
\name{extract_column}
\alias{extract_column}
\title{Extract specified column from the erp data and filter by the
specified category name}
\usage{
extract_column(df, colname)
}
\arguments{
\item{df}{tibble. The specified}

\item{cat_idx}{}
}
\description{
Extract specified column from the erp data and filter by the
specified category name
}
