% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-erp-spectrogram.R
\name{extract_specgram_mat}
\alias{extract_specgram_mat}
\title{Extracts the matrix of complex valued spectrogram entries}
\usage{
extract_specgram_mat(inp_sgram)
}
\arguments{
\item{inp_sgram}{The input spectrogram object as created from the
\code{signal::specgram} function}
}
\value{
a matrix of complex value spectrogram entries from the input
spectrogram object
}
\description{
Extracts the matrix of complex valued spectrogram entries
}
