% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-erp-create-matrices.R
\name{filter_rows}
\alias{filter_rows}
\title{Extract specified column from the erp data and filter by the
specified category name}
\usage{
filter_rows(df, cat_idx)
}
\arguments{
\item{df}{tibble. The specified}

\item{cat_idx}{}
}
\description{
Extract specified column from the erp data and filter by the
specified category name
}
