% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-read-erp.R
\name{get_document_name}
\alias{get_document_name}
\title{Parses out the document name from a filepath excluding the extension of
the document.
This is a helper function for the \code{analyzecorpus} function}
\usage{
get_document_name(document_path)
}
\arguments{
\item{document_path}{The filepath of a document or just the document name
(with or without extension)}
}
\value{
string containing the document name without the extension
of the document
}
\description{
Parses out the document name from a filepath excluding the extension of
the document.
This is a helper function for the \code{analyzecorpus} function
}
\examples{
doc_path <- "/Users/test/testing_doc.txt"
# The following function application returns "testing_doc"
get_document_name(doc_path)
}
