% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-read-erp.R
\name{get_erp_files}
\alias{get_erp_files}
\title{Used to generate a formatted tibble of erp files and filtering flags}
\usage{
get_erp_files(erp_dir)
}
\arguments{
\item{erp_dir}{The directory of erp files for a particular patient}
}
\value{
Tibble. formatted tibble of erp files and filtering flags
}
\description{
Used to generate a formatted tibble of erp files and filtering flags
}
