% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-erp-create-matrices.R
\name{get_fisher_abs_thrld}
\alias{get_fisher_abs_thrld}
\title{Obtain the 2-sided absolute Fisher information adjusted
threshold}
\usage{
get_fisher_abs_thrld(alph = 0.05, erp_filt_df)
}
\arguments{
\item{alph}{: overall alpha level (default value 0.05)}

\item{erp_filt_df}{: list of data frames containing filtered
ERP readings by channel, 1 data frame for each channel}
}
\description{
Obtain the 2-sided absolute Fisher information adjusted
threshold
}
