% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-read-erp.R
\name{read_erp_file}
\alias{read_erp_file}
\title{Read in the specified erp data files (non-labels)}
\usage{
read_erp_file(full_path, session_num, channel_num, filt_type, task_type)
}
\arguments{
\item{full_path}{The filepath to the specified erp label file}

\item{session_num}{The session number associated with each erp data file
i.e. typically {1, 2}}

\item{channel_num}{The channel number associated with each erp data file
i.e. typically {1, 2, ... 85}}

\item{filt_type}{The filter type associated with each erp data file
i.e. typically {"raw", "bandpass"}}

\item{task_type}{The task type associated with each erp data file
i.e. typically {"category", "individuation"}}
}
\value{
Tibble. The erp data file with session number, channel number,
filter type, task type and erp voltages for each millisecond X1:X1500
}
\description{
Read in the specified erp data files (non-labels)
}
