% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-read-erp.R
\name{read_erp_label_file}
\alias{read_erp_label_file}
\title{Read in the specified erp label files}
\usage{
read_erp_label_file(full_path, label_type)
}
\arguments{
\item{full_path}{The filepath to the specified erp label file}

\item{label_type}{The erp label type associated with the erp label file.
This will form the header for the erp file tibble once it is read
in}
}
\value{
Tibble. The erp label file with the erp label type as the header
}
\description{
Read in the specified erp label files
}
