% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-read-erp.R
\name{unrar_erp_file}
\alias{unrar_erp_file}
\title{Untar a tar.gz file to a specified output path}
\usage{
unrar_erp_file(targz_filepath, ext_path)
}
\arguments{
\item{targz_filepath}{The filepath to the tar.gz file}

\item{ext_path}{The filepath to extract the contents of the
\code{targz_filepath} file}
}
\description{
Untar a tar.gz file to a specified output path
}
